1、rescue和rescueDetail表单更新功能 2、rescueDetail表单病人基本信息从cache获取功能完善

+++ src/main/java/com/dhcc/cn/framework/controller/MedicalRescueController.java
@@ -111,7 +111,7 @@ public class MedicalRescueController {
     @PostMapping("/update")
     @ResponseResultBody
     @ResponseBody
-    public int update(MedicalRescueForm form){
+    public long update(MedicalRescueForm form){
         return mService.update(form);
     }
